requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
openpyxl>=3.1.0
xlrd>=2.0.0 